+++ src/OFS/Traversable.py
@@ -201,6 +201,10 @@ class Traversable:
         else:
             obj = self
 
+        # import time ordering problem
+        from webdav.NullResource import NullResource
+        resource = _marker
+
         try:
             while path:
                 name = path_pop()
@@ -295,6 +299,13 @@ class Traversable:
                             else:
                                 try:
                                     next = obj[name]
+                                    # The item lookup may return a
+                                    # NullResource, if this is the case we
+                                    # save it and return it if all other
+                                    # lookups fail.
+                                    if isinstance(next, NullResource):
+                                        resource = next
+                                        raise KeyError(name)
                                 except (AttributeError, TypeError):
                                     # Raise NotFound for easier debugging
                                     # instead of AttributeError: __getitem__
@@ -329,7 +340,11 @@ class Traversable:
                         except AttributeError:
                             raise e
                         if next is _marker:
-                            raise e
+                            # If we have a NullResource from earlier use it.
+                            next = resource
+                            if next is _marker:
+                                # Nothing found re-raise error
+                                raise e
 
                 obj = next