+++ src/ZPublisher/BaseRequest.py
@@ -18,6 +18,7 @@ from urllib.parse import quote as urllib_quote
 
 from AccessControl.ZopeSecurityPolicy import getRoles
 from Acquisition import aq_base
+from Acquisition import aq_inner
 from Acquisition.interfaces import IAcquirer
 from ExtensionClass import Base
 from zExceptions import Forbidden
@@ -179,6 +180,8 @@ class BaseRequest:
     The request object is a mapping object that represents a
     collection of variable to value mappings.
     """
+
+    maybe_webdav_client = 1
 
     # While the following assignment is not strictly necessary, it
     # prevents alot of unnecessary searches because, without it,
@@ -390,6 +393,9 @@ class BaseRequest:
             # index_html is still the default method, only any object can
             # override it by implementing its own __browser_default__ method
             method = 'index_html'
+        elif self.maybe_webdav_client:
+            # Probably a WebDAV client.
+            no_acquire_flag = 1
 
         URL = request['URL']
         parents = request['PARENTS']
@@ -450,6 +456,19 @@ class BaseRequest:
                     # BrowserDefault returns the object to be published
                     # (usually self) and a sequence of names to traverse to
                     # find the method to be published.
+
+                    # This is webdav support. The last object in the path
+                    # should not be acquired. Instead, a NullResource should
+                    # be given if it doesn't exist:
+                    if no_acquire_flag and \
+                       hasattr(object, 'aq_base') and \
+                       not hasattr(object, '__bobo_traverse__'):
+
+                        if (object.__parent__ is not
+                                aq_inner(object).__parent__):
+                            from webdav.NullResource import NullResource
+                            object = NullResource(parents[-2], object.getId(),
+                                                  self).__of__(parents[-2])
 
                     if IBrowserPublisher.providedBy(object):
                         adapter = object